PYIC-7578: Retry on SnapStartNotReadyException
We've seen some errors in perf tests where SnapStart not being ready has
caused the lambda invocation to fail. This happens when a function
hasn't been used for more than 14 days and the snapstart snapshot has
been binned by AWS. We can catch that error and retry to try to prevent
failures.

This doesn't go as far as retrying for other error types. Mostly because
we're not sure how idempotent our lambdas are, and we haven't seen a
problem so far so no need to fix it.

It was suggested we could try to retry the issues with are seeing when a
lambda segfaults. Unfortunatley the error returned to the step function
(`Runtime.ExitError`) isn't specific enought to reliably only retry that
error.

This also only fixes this issue for lambdas orchstrated by a
stepfunction. Those that are invoked directly by the API gateway will
suffer from this issue. APIGateway doesn't appear to offer anything, so
we'd need to handle and retry on the client side (core-front) if wanted
to try and fix them.