PYIC-7581: Simplify logging and reporting
This updates the logging of VC diffs to just show the signatures of the
VCs that are different between the tactical store and EVCS. This is
probably all we're interested in as we'll go and manually inspect those
VCs.

Also changes the reporting of signature validation failures to include
the CRI rather than the signature, which might be a bit more helpful.