PYIC-7581: Max out memory size
This sets the memory size for the reconciliation lambda to the maximum
allowed. This is done to increase the CPU available to it.

Having run tests, when checking signatures the lambda will process
100000 identities about 60% faster. The cost increase for this is about
1 cent per invocation, so probably worth it.

Also changes the default parallelism to the available CPUs. At 10240MB
memory we get 6 CPUs. The speed up isn't that significat compared to
using 4 (about 3%), but if we've got them we may as well use them.