PYIC-7571: Fetch all CRI by issuer once
When running the reconciliation lambda in parallel (more than one lambda
execution at once), we were seeing SSM rate limiting errors from
powertools. This is weird because it would appear that it's not always
respecting the cache.

To avoid this issue this create a new method on the config service which
will create a map of all configured issuers for the env to allow a
lookup of the associated CRI. We then use this to populate a map for the
script which it can refer to throughout the duration of its running.

I've run four parallel exections of this script with this change and
experienced no issues. This allowed the checking of the signatures for
100000 identities in less than two minutes.