PYIC-7580: Handle access-denied error response
Rather than a VC we may receive an asynchronous error response off the queue. In the case of a user abandon/abort (where the user has opted to abandon the app) the error response will contain an access-denied
At present when we process an asychronous response error, we simply update the CRI response item ('pending record') with status error regardless of the specific error code.
We should use a new status type abandon for this scenario and update the CRI response accordingly (with error used for all other error codes).