PYIC-7571: Track verifier use
Lots of VCs were failing signature verification in staging. Turns out
they were signed with an unexpected key. It would be useful to
understand how many VCs are signed with this key (which has been added
to config).
This would also potentially be interesting when we recondile in
production to understand the proportion of passport VCs signed with the
old and new keys.