PYIC-7632: Include kid in private key jwt for token exchange
When we exchange an auth code for a token with a CRI we use a signed JWT
to prove the request is coming from a legitimate source. This change
updates the header of the JWT to include the kid of the signing key.

We already had logic to include the key when signing - it was being used
when generating a JAR. It was implemented for DWP and is behind a flag
to allow it to be tested in staging before hitting prod.

This moves the logic to a shared place and utilises it for both signing
operations.