PYIC-7632: Refactor to improve reuseability
This refactors the signing logic so that we can share the same helpers
to do all the signing in both places that we sign JWTs.

It also ensures that the key used to create a signer is available with
the signer - this way we can be sure that the kid being used in the
header of a signed JWT is linked to the signer being used to sign.

This moves the kms signer lib we had into the common-services lib, as
they now depend on each other.