PYIC-7580: Handle access-denied for f2f in CheckExisitingIndentityHandler
Rather than a VC we may receive an asynchronous error response off the queue. In the case of a user abandon/abort (where the user has opted to abandon the app) the error response will contain an access-denied error code

F2F use access-denied but it seems they do. The behaviour will be the same (route to JOURNEY_F2F_FAIL) but we don’t want to warn log it as an unexpected status.