PYIC-7631: Refactor TokenRequestValidator
This refactors the validator to create a new clientAuthVerifier for
every request. This avoids having the Audience value set for the
lifecycle of a lambda. The objects are all pretty lightweight so
shouldn't be a problem.

Also tidies up a few other bits such as using instants. An
OffsetDateTime will change depending on the location of the system clock
- we should just be using UTC.