PYIC-7600: Add check-coi lambda to all IPV journeys
This adds the COI check to the evaluate subjourney so that we run it for
all successful IPV journeys.

It also removes the fallback to the tactical store if EVCS doesn't have
any VCs. All identities have now been migrated, and EVCS is the source
of truth. This simplifies things here as we don't have to worry about
partial identies being retrieved from tactical.