PYIC-7076: Add check-reverification-identity lambda
We need to ensure that if a user starts a reverification journey that
they have some form of existing identity to re-verify. This adds a new
lamdba that will check if they've got a P2, P1 or operational identity.
It's not concerned about CIs, expired VCs, account interventions or
pending identities.