PYIC-7761: Pass tracing headers to step function lambdas
The OpenTelemetry tracing framework being used by the lambdas needs the
WC3 headers to arrive in an expected format. It needs to be part of an
AWS event. This is a requirement coming from the lambda layer. It seems
to be very hard to supply them in any other way.

This updates the VTL in the openAPI spec to parse them from the http
request and include them in the input to the step function. It then
updates the step function definition to recreate the shape of an
APIGatewayProxyRequestEvent and supply it as input to the lambdas. The
lambdas are updated to expect a proxy event, and we then parse the usual
jourey/process/cri request objects from the body of the event.

This allows the tracing headers to be picked up early and propogated,
rather than generating new ones which breaks the traces in the UI.