PYIC-7076: Don't set targetVot for reverification journeys
It's not used in the journey as we don't actually do a gpg45 check, and
it makes no sense if the user has an operational profile.

This means that it's now possible for that value to be null in a users
session, so we handle that possibility where we read it.