PYIC-7761: Add OTEL tracing to AWS SDK and Java Http
This adds instrumentation to the AWS SDK and the Java HTTP client, so
that we can see more details traces in dynatrace.

This will be more useful on lambdas that are invoked directly by the API
gateway, as the trace ID from core-front will be propogated correctly,
allowing for a completely joined up view in dynatrace. For step-function
invoked lambdas the WC3 trace headers are not propegated correctly, and
so a new root span is geneated in the lambda. The AWS and http calls be
be traced with that new root as their parent. This may still be useful.