PYIC-7799: Create traffic when deploying
This updates the api-tests image to be able to run in a mode that will
just generate traffic in the build env. This is to ensure we have
traffic to inform the canarys.

It will poll for the status of the deploy action, and stop sending
traffic when it's no longer in progress.