PYIC-7811: Don't clear reprove flag from f2f response
As we're now only allowing a skip of starting a reprove identity journey
if the user has a pending identity in EVCS, we don't need to clear the
flag.

We only care if the user manages to achieve a profile - as we wouln't
want any further interventions to be skipped. And if they do achieve a
profile, then they will no longer have a pending identity. So won't
skip...

And if they don't achieve a profile, then they'll be on an identity
proving journey anyway.