PYIC-7746: Embedded metrics for CRI completion
This change starts sending custom metrics to track the rate of users
being sent to CRIs and then returning. We can compare these two metrics
to help us understand if the compeltion rates for different CRIs is
changing. And alarm on it in future.

We're using the logging mechanism, rather than sending custom metrics
directly with the cloudwatch SDK to avoid having to have a new AWS
client that needs to be initialised. The logs may also be useful for
more refined queries.

This uses the embedded metrics mechanism, rather than create a metric
log filter, as it's more explicit and will be harder to break
accidentally (by changing a log message for example).