PYIC-7799: Ensure TrafficGeneration continues on failure
If the API tests failed for any reason then the process returns a
non-zero exit code. Because we have the error exit option set, this
causes the background process to exit, and no more test cycles are run.

We can prevent this by falling back to returning true - this will ensure
that the loop continues.