PYIC-8058 Rethrow IOExceptions on EVCS requests
See P2 incident INC0014124

Currently we’re handling IOExceptions on EVCS requests and returning a graceful error response, rather than re-throwing wrapped in an unchecked exception to let the lambda crash.

In the last few weeks we’ve seen two instances of lambda check-existing-identity (one integration and one prod) spawned with apparently no network connectivity, raising IOExceptions on the HTTP request to EVCS. Every invocation of the instances failed with the same error.

For now let’s indiscriminately re-throw IOExceptions on EVCS calls, wrapped in a RuntimeException to let the lambda crash which will force a new instance to be spawned which should be healthy so subsequent requests are served without error.