PYIC-8223 Temporary mitigation to force lambda restart on fatal IOException
The previous mitigation (PYIC-8127) doesn't work in the way we expected - it turns
out throwing a runtime exception from the root lambda handler doesn't cause a
lambda instance to crash and restart (and we have confirmed this with AWS Support).

This takes things to the next level, calling System.exit() to force the program
to crash and restart the lambda instance (reproducible in dev). This is really
not nice but is a temporary mitigation while we get to the bottom of the root
cause of the issue, so that only a single invocation of an unhealthy lambda
instance fails before respawning.