PYIC-8724: Throw errors for failures
The manual-f2f-pending-reset Lambda currently shows all results in the same colour (green) in the interface, regardless of outcome. To reduce operator errors and align with the L1 user account deletion script, update the UI so that negative/error results are displayed in red instead of green. To do this we need to throw errors.