PYIC-8647 refine and correct api tests to validate sorry-technical flag
The old test coverage didn't include tests for server-errors on the cris
where the new functionality is introduced so a test for flag
disabled is needed. The tests for the new feature also
needed the flag activated in a different way than initailly implemented.
A default state for errors is needed in the yaml because the presence
of the feature flag check overrides the inherited default.