PYIC-8845: Add to path-ignore:       - '**/*.md'       - 'lambdas/*/src/test/**'       - 'libs/*/src/test/**'       - 'libs/test-data/**'       - 'libs/test-helpers/**'       - 'journey-map/**'       - 'local-running/**'
Add check-if-deploy-needed to trigger deployment if relevant files was changed.

Currently, the deployment of the API test image is tied to secure-post-merge.yml in GitHub Actions. We want to ensure that the API test image is deployed even when changes do not include the 'production' code.

Adding API tests to ignored paths is not feasible, as it would prevent the deployment of the test image when only those tests are modified.

Splitting the GitHub Actions workflow is also challenging, because we want the API test image to always be deployed before deploying the core-back application. Using the changed-files action provides a convenient solution to achieve this behaviour.